## BREAKING CHANGES:
- Rename `proto.Unmarshal` to `proto.UnmarshalValue` to make it clear what this function does. It is no longer receiving binary.ByteOrder interface, rather simply `arch byte` to select which byte order to use, 0 for little endian, otherwise big endian. This way, we avoid runtime interface lookup.
- Delete `proto.Marshal` and `proto.MarshalTo`, instead create method `(Value) MarshalAppend(b []byte, arch byte) ([]byte, error)`.
- Delete `WriteTo` method on `FileHeader`, `MessageDefinition` and `Message`, instead create method `(<Type>) MarshalAppend(b []byte) ([]byte, error)`. It will no longer handle allocations, it will only be responsible for appending bytes to given `b` (the responsibility is now shifted). This way, we avoid "abusing" sync.Pool for allocating a buffer to write to.
- Package `kit/byteorder` is now removed.

## What's Changed:
- Every instance of Encoder now have dedicated array of bytes `bytesArray` with the size of `proto.MaxBytesPerMessage` (130051 bytes)  to accomodate `MarshalAppend` before write it to the writer. This change has been made because the proto's marshaling behavior has been altered.

## Benchmarks

### Proto Marshal-Unmarshal
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/proto
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                 │   old.txt   │               new.txt               │
                 │   sec/op    │   sec/op     vs base                │
UnmarshalValue-4   6.149n ± 1%   4.616n ± 6%  -24.93% (p=0.000 n=10)

                 │  old.txt   │            new.txt             │
                 │    B/op    │    B/op     vs base            │
UnmarshalValue-4   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                 │  old.txt   │            new.txt             │
                 │ allocs/op  │ allocs/op   vs base            │
UnmarshalValue-4   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/proto
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                       │ oldmarshal.txt │           newmarshal.txt            │
                       │     sec/op     │   sec/op     vs base                │
MessageMarshalAppend-4     2.038µ ± 35%   1.228µ ± 4%  -39.76% (p=0.000 n=10)

                       │ oldmarshal.txt │         newmarshal.txt         │
                       │      B/op      │    B/op     vs base            │
MessageMarshalAppend-4       0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                       │ oldmarshal.txt │         newmarshal.txt         │
                       │   allocs/op    │ allocs/op   vs base            │
MessageMarshalAppend-4       0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

### Encode
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                                     │   old.txt    │               new.txt                │
                                     │    sec/op    │    sec/op     vs base                │
Encode/normal_header_zero-4            93.29m ±  9%   81.89m ±  1%  -12.23% (p=0.000 n=10)
Encode/normal_header_15-4              92.67m ± 53%   81.53m ±  1%  -12.02% (p=0.000 n=10)
Encode/compressed_timestamp_header-4   90.12m ± 16%   86.33m ± 34%        ~ (p=0.481 n=10)
geomean                                92.02m         83.22m         -9.56%

                                     │    old.txt    │              new.txt               │
                                     │     B/op      │    B/op     vs base                │
Encode/normal_header_zero-4            11149.0 ± 98%   182.0 ± 0%  -98.37% (p=0.000 n=10)
Encode/normal_header_15-4              11421.0 ± 96%   454.0 ± 0%  -96.02% (p=0.000 n=10)
Encode/compressed_timestamp_header-4   10304.0 ± 98%   181.0 ± 1%  -98.24% (p=0.000 n=10)
geomean                                10.69Ki         246.4       -97.75%

                                     │  old.txt   │               new.txt               │
                                     │ allocs/op  │ allocs/op   vs base                 │
Encode/normal_header_zero-4            4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
Encode/normal_header_15-4              14.00 ± 0%   14.00 ± 0%       ~ (p=1.000 n=10) ¹
Encode/compressed_timestamp_header-4   4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                6.073        6.073       +0.00%
¹ all samples are equal
```

### Decode
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │   old.txt    │              new.txt               │
         │    sec/op    │   sec/op     vs base               │
Decode-4   103.93m ± 9%   94.25m ± 3%  -9.31% (p=0.000 n=10)

         │   old.txt    │            new.txt             │
         │     B/op     │     B/op      vs base          │
Decode-4   73.49Mi ± 0%   73.49Mi ± 0%  ~ (p=0.935 n=10)

         │   old.txt   │             new.txt             │
         │  allocs/op  │  allocs/op   vs base            │
Decode-4   100.0k ± 0%   100.0k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

```